Improve blob storage of mutation/subscription prefixes
