+++ tests/normalization-tests/src/commonTest/kotlin/com/example/NormalizationTest.kt
@@ -6,7 +6,6 @@ import com.apollographql.apollo.api.json.jsonReader
 import com.apollographql.apollo.api.toApolloResponse
 import com.apollographql.cache.normalized.ApolloStore
 import com.apollographql.cache.normalized.FetchPolicy
-import com.apollographql.cache.normalized.allRecords
 import com.apollographql.cache.normalized.api.CacheKey
 import com.apollographql.cache.normalized.api.CacheKeyGenerator
 import com.apollographql.cache.normalized.api.CacheKeyGeneratorContext
@@ -15,17 +14,13 @@ import com.apollographql.cache.normalized.api.CacheResolver
 import com.apollographql.cache.normalized.api.FieldPolicyCacheResolver
 import com.apollographql.cache.normalized.api.ResolverContext
 import com.apollographql.cache.normalized.api.TypePolicyCacheKeyGenerator
-import com.apollographql.cache.normalized.apolloStore
 import com.apollographql.cache.normalized.fetchPolicy
 import com.apollographql.cache.normalized.memory.MemoryCacheFactory
 import com.apollographql.cache.normalized.normalizedCache
-import com.apollographql.cache.normalized.sql.SqlNormalizedCacheFactory
 import com.apollographql.cache.normalized.store
-import com.apollographql.cache.normalized.testing.append
 import com.apollographql.cache.normalized.testing.runTest
 import com.apollographql.mockserver.MockServer
 import com.apollographql.mockserver.enqueueString
-import com.example.four.CreateUserMutation
 import com.example.one.Issue2818Query
 import com.example.one.Issue3672Query
 import com.example.one.fragment.SectionFragment
@@ -40,7 +35,6 @@ import com.example.two.GetCountryQuery
 import com.example.two.NestedFragmentQuery
 import okio.Buffer
 import kotlin.test.Test
-import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
 
 internal object IdBasedCacheKeyResolver : CacheResolver, CacheKeyGenerator {
@@ -311,60 +305,6 @@ class NormalizationTest {
     val fromCache2 =
       apolloClient.query(GetBooksByIdsPaginatedNoCursorsWithFragmentQuery(listOf("book-1"))).fetchPolicy(FetchPolicy.CacheOnly).execute()
     assertEquals("First book", fromCache2.data?.viewer?.libraries?.first()?.booksPaginated?.edges?.first()?.bookEdge?.node?.name)
-
-    apolloClient.close()
-    mockserver.close()
-  }
-
-  @Test
-  fun mutationRoot() = runTest {
-    val mockserver = MockServer()
-    val apolloClient = ApolloClient.Builder()
-        .serverUrl(mockserver.url())
-        .normalizedCache(SqlNormalizedCacheFactory())
-        .build()
-
-    apolloClient.apolloStore.clearAll()
-    mockserver.enqueueString(
-        // language=JSON
-        """
-        {
-          "data": {
-            "createUser": {
-              "__typename": "User",
-              "id": "user-1",
-              "name": "John Doe",
-              "projects": [
-                {
-                  "__typename": "Project",
-                  "id": "project-1",
-                  "name": "Project 1",
-                  "description": "Description 1",
-                  "owner": {
-                    "__typename": "User",
-                    "id": "user-2",
-                    "name": "Jane Doe"
-                  }
-                }
-              ]
-            }
-          }
-        }
-        """.trimIndent()
-    )
-    apolloClient.mutation(CreateUserMutation("John")).fetchPolicy(FetchPolicy.NetworkOnly).execute()
-
-    apolloClient.apolloStore.accessCache { normalizedCache ->
-      assertContentEquals(
-          listOf(
-              CacheKey.MUTATION_ROOT,
-              CacheKey.MUTATION_ROOT.append("""createUser({"name":"John"})"""),
-              CacheKey.MUTATION_ROOT.append("""createUser({"name":"John"})""", "projects", "0"),
-              CacheKey.MUTATION_ROOT.append("""createUser({"name":"John"})""", "projects", "0", "owner"),
-          ),
-          normalizedCache.allRecords().keys,
-      )
-    }
 
     apolloClient.close()
     mockserver.close()
